Capitalize setting and moved function definition to Setting.cpp
