+++ src/Utils/Setting.cpp
@@ -1,0 +1,12 @@
+#include "Setting.h"
+
+Setting::Setting(QObject *parent)
+    : QObject{parent}
+{
+    bool isElysia() {return isElysia_;}
+
+    void setIsElysia(bool value) {
+        isElysia_ = value;
+        emit isElysiaChanged(isElysia_);
+    };
+}
