--- src/Utils/setting.cpp
@@ -1,7 +1,0 @@
-#include "setting.h"
-
-Setting::Setting(QObject *parent)
-    : QObject{parent}
-{
-
-}
